% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{dr_running}
\alias{dr_running}
\title{Is the drogonR server currently running?}
\usage{
dr_running()
}
\value{
`TRUE` if a server is running in this process, `FALSE` otherwise.
}
\description{
Is the drogonR server currently running?
}
