% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drogonR-package.R
\docType{package}
\name{drogonR-package}
\alias{drogonR}
\alias{drogonR-package}
\title{drogonR: High-Performance HTTP Server for R via Drogon}
\description{
Provides an R interface to the Drogon C++ HTTP server framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Zabis13/drogonR}
  \item Report bugs at \url{https://github.com/Zabis13/drogonR/issues}
}

}
\author{
\strong{Maintainer}: Yuri Baramykov \email{lbsbmsu@mail.ru} (\href{https://orcid.org/0009-0000-7627-4217}{ORCID})

Other contributors:
\itemize{
  \item An Tao (Author of the bundled Drogon and Trantor C++ libraries) [contributor, copyright holder]
  \item Shuo Chen (Author of the Muduo library, on which Trantor is based) [contributor, copyright holder]
  \item Baptiste Lepilleur (Original author of the bundled JsonCpp library) [contributor, copyright holder]
  \item Christopher Dunn (Maintainer of JsonCpp) [contributor]
  \item  JsonCpp Contributors (See src/drogon/third_party/jsoncpp/AUTHORS in the package source) [contributor, copyright holder]
}

}
\keyword{internal}
